+++ bugsnag-spring/javax/src/main/java/com/bugsnag/SpringBootJavaxConfiguration.java
@@ -17,11 +17,11 @@ import javax.servlet.ServletRequestListener;
 class SpringBootJavaxConfiguration extends SpringBootConfiguration {
 
     /**
-    * The {@link com.bugsnag.servlet.javax.BugsnagServletContainerInitializer} does not work for Spring Boot, need to
-    * register the {@link BugsnagServletRequestListener} using a Spring Boot
-    * {@link ServletListenerRegistrationBean} instead. This adds session tracking and
-    * automatic servlet request metadata collection.
-    */
+     * The {@link com.bugsnag.servlet.javax.BugsnagServletContainerInitializer} does not work for Spring Boot, need to
+     * register the {@link BugsnagServletRequestListener} using a Spring Boot
+     * {@link ServletListenerRegistrationBean} instead. This adds session tracking and
+     * automatic servlet request metadata collection.
+     */
     @Bean
     @Conditional(SpringWebMvcLoadedCondition.class)
     ServletListenerRegistrationBean<ServletRequestListener> listenerRegistrationBean() {
