+++ bugsnag/src/main/java/com/bugsnag/servlet/BugsnagServletContainerInitializer.java
@@ -1,0 +1,9 @@
+package com.bugsnag.servlet;
+
+/**
+ * @see com.bugsnag.servlet.javax.BugsnagServletContainerInitializer
+ * @deprecated since 3.7.1 - to be replaced with {@code com.bugsnag.servlet.javax.BugsnagServletContainerInitializer}
+ */
+@Deprecated
+public class BugsnagServletContainerInitializer extends com.bugsnag.servlet.javax.BugsnagServletContainerInitializer {
+}
