+++ bugsnag/src/main/java/com/bugsnag/servlet/BugsnagServletRequestListener.java
@@ -1,0 +1,9 @@
+package com.bugsnag.servlet;
+
+/**
+ * @see com.bugsnag.servlet.javax.BugsnagServletRequestListener
+ * @deprecated since 3.7.1 - to be replaced with {@code com.bugsnag.servlet.javax.BugsnagServletRequestListener}
+ */
+@Deprecated
+public class BugsnagServletRequestListener extends com.bugsnag.servlet.javax.BugsnagServletRequestListener {
+}
