+++ bugsnag/src/main/resources/META-INF/services/javax.servlet.ServletContainerInitializer
@@ -1,1 +1,1 @@
-com.bugsnag.servlet.javax.BugsnagServletContainerInitializer
+com.bugsnag.servlet.BugsnagServletContainerInitializer
